Remove old examples and reorganize package layout
